Fix cross-platform build: restore uint64 conversion for stat.Dev
stat.Dev type varies by platform (int32 on some Unix systems, uint64 on
others). The conversion is required for portability. Added nolint directive
to suppress unconvert warning on platforms where it's already uint64.

+++ internal/scanner/device_unix.go
@@ -13,5 +13,6 @@ func getDeviceID(info os.FileInfo) (uint64, bool) {
 	if !ok {
 		return 0, false
 	}
-	return stat.Dev, true
+	//nolint:unconvert // stat.Dev type varies by platform (int32 on some, uint64 on others)
+	return uint64(stat.Dev), true
 }
